chore: split extras and variant features into modules
