+++ js/proficiencies.js
@@ -1,0 +1,49 @@
+export const ARTISAN_TOOLS = [
+  "Alchemist's Supplies",
+  "Brewer's Supplies",
+  "Calligrapher's Tools",
+  "Carpenter's Tools",
+  "Cartographer's Tools",
+  "Cobbler's Tools",
+  "Cook's Utensils",
+  "Glassblower's Tools",
+  "Jeweler's Tools",
+  "Leatherworker's Tools",
+  "Mason's Tools",
+  "Painter's Supplies",
+  "Potter's Tools",
+  "Smith's Tools",
+  "Tinker's Tools",
+  "Weaver's Tools",
+  "Woodcarver's Tools",
+];
+
+export const MUSICAL_INSTRUMENTS = [
+  "Bagpipes",
+  "Drum",
+  "Dulcimer",
+  "Flute",
+  "Lute",
+  "Lyre",
+  "Horn",
+  "Pan Flute",
+  "Shawm",
+  "Viol",
+];
+
+export const ALL_TOOLS = [
+  ...ARTISAN_TOOLS,
+  ...MUSICAL_INSTRUMENTS,
+  "Disguise Kit",
+  "Forgery Kit",
+  "Herbalism Kit",
+  "Navigator's Tools",
+  "Poisoner's Kit",
+  "Thieves' Tools",
+  "Dice Set",
+  "Dragonchess Set",
+  "Playing Card Set",
+  "Three-Dragon Ante Set",
+  "Vehicle (Land)",
+  "Vehicle (Water)",
+];
