+++ js/script.js
@@ -3,405 +3,25 @@ import { loadSpells, filterSpells, handleSpellcasting } from './spellcasting.js'
 import { convertRaceData, ALL_SKILLS } from './raceData.js';
 import { getSelectedData, setSelectedData, saveSelectedData } from './state.js';
 import { createHeader, createParagraph, createList } from './domHelpers.js';
+import { ARTISAN_TOOLS, MUSICAL_INSTRUMENTS, ALL_TOOLS } from './proficiencies.js';
+import { updateVariantSkillOptions, handleVariantExtraSelections, handleVariantFeatureChoices } from './variantFeatures.js';
+import { handleExtraLanguages, handleExtraSkills, handleExtraTools, handleExtraAncestry, gatherRaceTraitSelections } from './extrasSelections.js';
+import { openExtrasModal, updateExtraSelectionsView, showExtraSelection, extraCategoryAliases, extraCategoryDescriptions } from './extrasModal.js';
 
-// ==================== MAPPING PER LE EXTRA VARIANT FEATURES ====================
-const variantExtraMapping = {
-  "Drow Magic": {
-    type: "none" // Le spell fisse saranno gestite separatamente
-  },
-  "Skill Versatility": {
-    type: "skills",
-    count: 2,
-    options: [
-      "Acrobatics", "Animal Handling", "Arcana", "Athletics", "Deception", "History",
-      "Insight", "Intimidation", "Investigation", "Medicine", "Nature", "Perception",
-      "Performance", "Persuasion", "Religion", "Sleight of Hand", "Stealth", "Survival"
-    ]
-  },
-  "Swim": {
-    type: "none"
-  },
-  "Cantrip (Wizard)": {
-    type: "spells",
-    filter: "level=0|class=Wizard"
-  }
-  // Aggiungi altri mapping se necessario
-};
-
-// ==================== LISTE DI STRUMENTI ====================
-const ARTISAN_TOOLS = [
-  "Alchemist's Supplies",
-  "Brewer's Supplies",
-  "Calligrapher's Tools",
-  "Carpenter's Tools",
-  "Cartographer's Tools",
-  "Cobbler's Tools",
-  "Cook's Utensils",
-  "Glassblower's Tools",
-  "Jeweler's Tools",
-  "Leatherworker's Tools",
-  "Mason's Tools",
-  "Painter's Supplies",
-  "Potter's Tools",
-  "Smith's Tools",
-  "Tinker's Tools",
-  "Weaver's Tools",
-  "Woodcarver's Tools"
-];
-
-const MUSICAL_INSTRUMENTS = [
-  "Bagpipes",
-  "Drum",
-  "Dulcimer",
-  "Flute",
-  "Lute",
-  "Lyre",
-  "Horn",
-  "Pan Flute",
-  "Shawm",
-  "Viol"
-];
-
-export const ALL_TOOLS = [
-  ...ARTISAN_TOOLS,
-  ...MUSICAL_INSTRUMENTS,
-  "Disguise Kit",
-  "Forgery Kit",
-  "Herbalism Kit",
-  "Navigator's Tools",
-  "Poisoner's Kit",
-  "Thieves' Tools",
-  "Dice Set",
-  "Dragonchess Set",
-  "Playing Card Set",
-  "Three-Dragon Ante Set",
-  "Vehicle (Land)",
-  "Vehicle (Water)",
-];
-
-// ==================== FUNZIONI PER LE VARIANT FEATURES ====================
-function updateVariantSkillOptions() {
-  const allVariantSkillSelects = document.querySelectorAll(".variantSkillChoice");
-  if (!allVariantSkillSelects.length) return;
-
-  const selected = new Set([...allVariantSkillSelects].map(select => select.value).filter(Boolean));
-
-  allVariantSkillSelects.forEach(select => {
-    const current = select.value;
-    select.innerHTML = `<option value="">Seleziona...</option>`;
-    JSON.parse(select.getAttribute("data-options")).forEach(skill => {
-      const option = document.createElement("option");
-      option.value = skill;
-      option.textContent = skill;
-      if (skill === current || !selected.has(skill)) {
-        select.appendChild(option);
-        if (skill === current) option.selected = true;
-      }
-    });
-  });
+export let availableLanguages = [];
+export function setAvailableLanguages(langs) {
+  availableLanguages = langs;
 }
 
-function handleVariantExtraSelections() {
-  const variantElem = document.getElementById("variantFeatureChoice");
-  const container = document.getElementById("variantExtraContainer");
-  if (!container || !variantElem || !variantElem.value) return;
-  
-  const selectedVariant = variantElem.value;
-  const mapData = variantExtraMapping[selectedVariant];
+let selectedData = getSelectedData();
 
-  container.innerHTML = ""; // Pulisce il contenuto precedente
-
-  if (!mapData) return;
-
-  if (mapData.type === "skills") {
-    container.innerHTML = `<p><strong>Seleziona ${mapData.count} skill per ${selectedVariant}:</strong></p>` +
-      Array(mapData.count).fill(0).map((_, i) =>
-        `<select class="variantSkillChoice" id="variantSkillChoice${i}" data-options='${JSON.stringify(mapData.options)}' onchange="updateVariantSkillOptions()">
-            <option value="">Seleziona...</option>
-            ${mapData.options.map(s => `<option value="${s}">${s}</option>`).join("")}
-          </select>`
-      ).join(" ");
-    container.querySelectorAll("select").forEach(sel => {
-      sel.addEventListener("change", () => {
-        selectedData["Variant Feature Skills"] = [...container.querySelectorAll(".variantSkillChoice")]
-          .map(s => s.value)
-          .filter(Boolean);
-        checkTraitCompletion("variantFeatureTrait");
-      });
-    });
-    checkTraitCompletion("variantFeatureTrait");
-  } else if (mapData.type === "spells") {
-    loadSpells(spellList => {
-      const filtered = filterSpells(spellList, mapData.filter);
-
-      container.innerHTML = filtered.length
-        ? `<p><strong>Seleziona un incantesimo per ${selectedVariant}:</strong></p>
-            <select id="variantSpellChoice">
-              <option value="">Seleziona...</option>
-              ${filtered.map(spell => `<option value="${spell.name}">${spell.name}</option>`).join("")}
-            </select>`
-        : `<p>Nessun incantesimo trovato per il filtro: ${mapData.filter}</p>`;
-
-      const spellSel = document.getElementById("variantSpellChoice");
-      if (spellSel) {
-        spellSel.addEventListener("change", () => {
-          selectedData["Variant Feature Spell"] = [spellSel.value];
-          checkTraitCompletion("variantFeatureTrait");
-        });
-        checkTraitCompletion("variantFeatureTrait");
-      }
-    });
-  }
-}
-
-function handleVariantFeatureChoices(data) {
-  if (!data.variant_feature_choices) return;
-
-  console.log(`📌 Trovata Variant Feature per ${data.name}:`, data.variant_feature_choices);
-
-  const container = document.getElementById("variantFeatureSelectionContainer");
-  if (!container) return;
-
-  let html = `<p><strong>Scegli una Variant Feature:</strong></p><select id="variantFeatureChoice">
-                <option value="">Seleziona...</option>`;
-  data.variant_feature_choices.forEach(opt => {
-    html += `<option value="${opt.name}">${opt.name}</option>`;
-  });
-  html += `</select>`;
-
-  container.innerHTML = html;
-  document.getElementById("variantFeatureChoice").addEventListener("change", handleVariantExtraSelections);
-}
-
-// ==================== EXTRAS: LANGUAGES, SKILLS, TOOLS, ANCESTRY ====================
-function handleExtraLanguages(data, containerId) {
-  if (data.languages && data.languages.choice > 0) {
-    loadLanguages(langs => {
-      const availableLangs = langs.filter(lang => !data.languages.fixed.includes(lang));
-      const options = availableLangs.map(lang => `<option value="${lang}">${lang}</option>`).join("");
-      const html = `<h4>Lingue Extra</h4>
-                    <select id="extraLanguageDropdown">
-                      <option value="">Seleziona...</option>
-                      ${options}
-                    </select>`;
-      const container = document.getElementById(containerId);
-      if (container) container.innerHTML = html;
-    });
-  } else {
-    const container = document.getElementById(containerId);
-    if (container) container.innerHTML = "";
-  }
-}
-
-function handleExtraSkills(data, containerId) {
-  if (data.skill_choices) {
-    const skillContainer = document.createElement("div");
-    const title = document.createElement("h4");
-    title.textContent = "Skill Extra";
-    skillContainer.appendChild(title);
-
-    const selects = [];
-    for (let i = 0; i < data.skill_choices.number; i++) {
-      const select = document.createElement("select");
-      select.classList.add("skillChoice");
-      select.id = `skillChoice${i}`;
-      select.dataset.options = JSON.stringify(data.skill_choices.options);
-
-      const defaultOption = document.createElement("option");
-      defaultOption.value = "";
-      defaultOption.textContent = "Seleziona...";
-      select.appendChild(defaultOption);
-
-      data.skill_choices.options.forEach(skill => {
-        const option = document.createElement("option");
-        option.value = skill;
-        option.textContent = skill;
-        select.appendChild(option);
-      });
-
-      skillContainer.appendChild(select);
-      selects.push(select);
-    }
-
-    const update = () => {
-      const chosen = new Set(selects.map(s => s.value).filter(Boolean));
-      selects.forEach(sel => {
-        const opts = JSON.parse(sel.dataset.options);
-        const current = sel.value;
-        sel.innerHTML = `<option value="">Seleziona...</option>` +
-          opts.map(o => `<option value="${o}" ${chosen.has(o) && o !== current ? 'disabled' : ''}>${o}</option>`).join('');
-        sel.value = current;
-      });
-    };
-    selects.forEach(sel => sel.addEventListener('change', update));
-    update();
-
-    const container = document.getElementById(containerId);
-    if (container) container.appendChild(skillContainer);
-  }
-}
-
-function handleExtraTools(data, containerId) {
-  if (data.tool_choices) {
-    const container = document.getElementById(containerId);
-    const toolDiv = document.createElement('div');
-    const title = document.createElement('h4');
-    title.textContent = 'Tool Extra';
-    toolDiv.appendChild(title);
-
-    const selects = [];
-    for (let i = 0; i < data.tool_choices.number; i++) {
-      const select = document.createElement('select');
-      select.classList.add('toolChoice');
-      select.id = `toolChoice${i}`;
-      select.dataset.options = JSON.stringify(data.tool_choices.options);
-
-      const defaultOption = document.createElement('option');
-      defaultOption.value = '';
-      defaultOption.textContent = 'Seleziona...';
-      select.appendChild(defaultOption);
-
-      data.tool_choices.options.forEach(tool => {
-        const option = document.createElement('option');
-        option.value = tool;
-        option.textContent = tool;
-        select.appendChild(option);
-      });
-
-      toolDiv.appendChild(select);
-      selects.push(select);
-    }
-
-    const update = () => {
-      const chosen = new Set(selects.map(s => s.value).filter(Boolean));
-      selects.forEach(sel => {
-        const opts = JSON.parse(sel.dataset.options);
-        const current = sel.value;
-        sel.innerHTML = `<option value="">Seleziona...</option>` +
-          opts.map(o => `<option value="${o}" ${chosen.has(o) && o !== current ? 'disabled' : ''}>${o}</option>`).join('');
-        sel.value = current;
-      });
-    };
-    selects.forEach(sel => sel.addEventListener('change', update));
-    update();
-
-    if (container) {
-      container.innerHTML = '';
-      container.appendChild(toolDiv);
-    }
-  } else {
-    const container = document.getElementById(containerId);
-    if (container) container.innerHTML = '';
-  }
-}
-
-function handleExtraAncestry(data, containerId) {
-  if (data.variant_feature_choices && data.variant_feature_choices.length > 0) {
-    const ancestryOptions = data.variant_feature_choices.filter(opt => opt.name.toLowerCase().includes("ancestry"));
-    if (ancestryOptions.length > 0) {
-      let html = `<h4>Seleziona Ancestry</h4>
-                  <select id="ancestrySelection">
-                    <option value="">Seleziona...</option>`;
-      ancestryOptions.forEach(opt => {
-        html += `<option value="${opt.name}">${opt.name}</option>`;
-      });
-      html += `</select>`;
-      const container = document.getElementById(containerId);
-      if (container) container.innerHTML = html;
-    }
-  } else {
-    const container = document.getElementById(containerId);
-    if (container) container.innerHTML = "";
-  }
-}
-
-function checkTraitCompletion(detailId) {
+export function checkTraitCompletion(detailId) {
   const detail = document.getElementById(detailId);
   if (!detail) return;
   const selects = detail.querySelectorAll("select");
   const incomplete = [...selects].some(sel => !sel.value);
   detail.classList.toggle("incomplete", incomplete);
 }
-
-function gatherRaceTraitSelections() {
-  const result = {};
-  const lang = document.getElementById("extraLanguageDropdown")?.value;
-  if (lang) result.languages = [lang];
-
-  const skills = [...document.querySelectorAll("#skillSelectionContainer .skillChoice")]
-    .map(s => s.value)
-    .filter(Boolean);
-  if (skills.length) result.skills = skills;
-
-  const tools = [...document.querySelectorAll("#toolSelectionContainer .toolChoice")]
-    .map(s => s.value)
-    .filter(Boolean);
-  if (tools.length) result.tools = tools;
-
-  const variant = document.getElementById("variantFeatureChoice")?.value;
-  if (variant) {
-    result.variantFeature = variant;
-    const vSkills = [...document.querySelectorAll(".variantSkillChoice")]
-      .map(s => s.value)
-      .filter(Boolean);
-    if (vSkills.length) result.variantSkills = vSkills;
-    const vSpell = document.getElementById("variantSpellChoice")?.value;
-    if (vSpell) result.variantSpell = vSpell;
-  }
-
-  const ancestry = document.getElementById("ancestrySelection")?.value;
-  if (ancestry) result.ancestry = ancestry;
-
-  return result;
-}
-
-// ==================== POPUP FOR EXTRA SELECTIONS ====================
-
-// Global variables for the extra selections popup
-let selectedData = getSelectedData();
-
-// These variables are referenced throughout the extra selections workflow.
-// In earlier refactors they were implicitly assumed to exist which caused
-// runtime ReferenceError issues when the script loaded before any
-// selections were defined.  Initialising them here ensures that later
-// checks like `extraSelections.every(...)` can safely run even if no
-// extra selections have been generated yet.
-let extraSelections = [];
-let currentSelectionIndex = 0;
-let extraModalContext = "";
-
-// Elements used to navigate and close the extra selections modal
-const prevTraitEl = document.getElementById("prevTrait");
-const nextTraitEl = document.getElementById("nextTrait");
-const closeModalEl = document.getElementById("closeModal");
-
-// Cached list of all languages loaded from JSON
-export let availableLanguages = [];
-export function setAvailableLanguages(langs) {
-  availableLanguages = langs;
-}
-// Flag to track confirmation of class selection
-
-// Mapping and descriptions for extra selection categories
-const extraCategoryAliases = {
-  "Cantrip": "Cantrips",
-  "Cantrips": "Cantrips",
-  "Skill Proficiency": "Skill Proficiency",
-  "Tool Proficiency": "Tool Proficiency",
-  "Fighting Style": "Fighting Style",
-  "Additional Fighting Style": "Fighting Style",
-  "Divine Domain": "Divine Domain",
-  "Metamagic": "Metamagic"
-};
-
-const extraCategoryDescriptions = {
-  "Cantrips": "Scegli i tuoi cantrip.",
-  "Skill Proficiency": "Seleziona le competenze nelle abilità.",
-  "Tool Proficiency": "Seleziona le competenze negli strumenti.",
-  "Fighting Style": "Scegli il tuo stile di combattimento.",
-  "Divine Domain": "Seleziona il tuo dominio divino.",
-  "Metamagic": "Scegli le opzioni di Metamagia."
-};
 
 /**
  * Returns a merged list of proficiencies already taken via
@@ -696,239 +316,6 @@ function initializeAccordion(root) {
 /**
  * Render extra selections inside an accordion rather than a modal.
  */
-function openExtrasModal(selections, context = "race") {
-  if (!selections || selections.length === 0) {
-    console.warn("⚠️ Nessuna selezione extra disponibile, il pop-up non verrà mostrato.");
-    return;
-  }
-
-  // Store the selections and context so that navigation and other
-  // handlers (e.g. the close button) can reference them safely.
-  extraSelections = selections;
-  currentSelectionIndex = 0;
-  extraModalContext = context;
-
-  const containerId = context === "class" ? "classExtrasAccordion" : "raceExtraTraitsContainer";
-  const container = document.getElementById(containerId);
-  if (!container) return;
-  container.classList.remove('hidden');
-
-  // Ensure categories exist in selectedData
-  selections.forEach(selection => {
-    const key = extraCategoryAliases[selection.name] || selection.name;
-    if (!selectedData[key]) {
-      selectedData[key] = [];
-    }
-  });
-
-  container.innerHTML = "";
-
-  selections.forEach((selection, selIdx) => {
-    const categoryKey = extraCategoryAliases[selection.name] || selection.name;
-    const item = document.createElement('div');
-    item.classList.add('accordion-item', 'user-choice');
-
-    const header = document.createElement('button');
-    header.type = 'button';
-    header.classList.add('accordion-header');
-    header.textContent = selection.name;
-    item.appendChild(header);
-
-    const content = document.createElement('div');
-    content.classList.add('accordion-content');
-
-    for (let i = 0; i < selection.count; i++) {
-      const sel = document.createElement('select');
-      sel.classList.add('extra-selection');
-      sel.dataset.category = categoryKey;
-      sel.dataset.index = i;
-      sel.innerHTML = `<option value="">Seleziona...</option>` +
-        selection.selection.map(opt => `<option value="${opt}">${opt}</option>`).join("");
-      content.appendChild(sel);
-
-      sel.addEventListener('change', e => {
-        const category = e.target.dataset.category;
-        const index = e.target.dataset.index;
-        if (!selectedData[category]) {
-          selectedData[category] = [];
-        }
-        selectedData[category][index] = e.target.value;
-        saveSelectedData();
-        updateExtraSelectionsView();
-      });
-    }
-
-    item.appendChild(content);
-    container.appendChild(item);
-  });
-
-  initializeAccordion(container);
-  updateExtraSelectionsView();
-}
-
-function updateExtraSelectionsView() {
-  console.log("🔄 Recupero selezioni extra salvate...");
-
-  // ✅ Recupera i dati attuali dallo stato
-  selectedData = getSelectedData();
-
-  function updateContainer(id, title, dataKey) {
-    const container = document.getElementById(id);
-    if (!container) return;
-
-    const values = (selectedData[dataKey] || []).filter(v => v);
-
-    if (values.length > 0) {
-      container.innerHTML = `<p><strong>${title}:</strong> ${values.join(", ")}</p>`;
-      container.classList.remove('hidden');
-    } else {
-      container.innerHTML = "";
-      container.classList.add('hidden');
-    }
-  }
-
-  const summaryMap = extraModalContext === "class" ? {} : {
-    "Languages": ["languageSelection", "Lingue Extra"],
-    "Skill Proficiency": ["skillSelectionContainer", "Skill Proficiency"],
-    "Tool Proficiency": ["toolSelectionContainer", "Tool Proficiency"]
-    // Spellcasting selections are now integrated directly within trait details
-  };
-  Object.entries(summaryMap).forEach(([key, [id, title]]) => {
-    if (selectedData[key] !== undefined) {
-      updateContainer(id, title, key);
-    } else {
-      const container = document.getElementById(id);
-      if (container) container.classList.add('hidden');
-    }
-  });
-
-  console.log("✅ Extra selections aggiornate:", selectedData);
-}
-
-/**
- * Displays the current extra selection in the popup.
- * Each dropdown gets a data-category attribute set to the current selection's name.
- * The "Close" button is shown only when on the last extra selection.
- */
-function showExtraSelection() {
-  const titleElem = document.getElementById("extraTraitTitle");
-  const descElem = document.getElementById("extraTraitDescription");
-  const selectionElem = document.getElementById("extraTraitSelection");
-
-  if (!extraSelections || extraSelections.length === 0) return;
-
-  const currentSelection = extraSelections[currentSelectionIndex];
-
-  titleElem.innerText = currentSelection.name;
-  const desc = currentSelection.description || extraCategoryDescriptions[currentSelection.name] || "";
-  descElem.innerText = desc;
-  selectionElem.innerHTML = ""; // Pulisce il contenuto precedente
-
-  if (currentSelection.selection) {
-    const categoryKey = extraCategoryAliases[currentSelection.name] || currentSelection.name;
-    const typeLookup = {
-      Languages: 'languages',
-      'Skill Proficiency': 'skills',
-      'Tool Proficiency': 'tools',
-    };
-    const taken = new Set(getTakenProficiencies(typeLookup[categoryKey] || ''));
-    const selectedValues = new Set((selectedData[categoryKey] || []).filter(v => v));
-    taken.forEach(v => selectedValues.add(v));
-
-    let dropdownHTML = "";
-    for (let i = 0; i < currentSelection.count; i++) {
-      dropdownHTML += `<select class="extra-selection" data-category="${categoryKey}" data-index="${i}">
-                        <option value="">Seleziona...</option>`;
-      currentSelection.selection.forEach(option => {
-        const disabled = selectedValues.has(option) && !selectedData[categoryKey]?.includes(option);
-        dropdownHTML += `<option value="${option}" ${disabled ? "disabled" : ""}>${option}</option>`;
-      });
-      dropdownHTML += `</select><br>`;
-    }
-    selectionElem.innerHTML = dropdownHTML;
-
-    document.querySelectorAll(".extra-selection").forEach(select => {
-      select.addEventListener("change", (event) => {
-        const rawCategory = event.target.getAttribute("data-category");
-        const category = extraCategoryAliases[rawCategory] || rawCategory;
-        const index = event.target.getAttribute("data-index");
-
-        if (!selectedData[category]) {
-          selectedData[category] = [];
-        }
-
-        selectedData[category][index] = event.target.value;
-
-        console.log(`📝 Salvato: ${category} -> ${selectedData[category]}`);
-
-        saveSelectedData();
-        updateExtraSelectionsView();
-      });
-    });
-  }
-
-  if (prevTraitEl && nextTraitEl && closeModalEl) {
-    prevTraitEl.disabled = (currentSelectionIndex === 0);
-    nextTraitEl.disabled = (currentSelectionIndex === extraSelections.length - 1);
-
-    const allChoicesFilled = extraSelections.every(sel =>
-      selectedData[sel.name] && selectedData[sel.name].filter(v => v).length === sel.count
-    );
-
-    if (currentSelectionIndex === extraSelections.length - 1 && allChoicesFilled) {
-      closeModalEl.style.display = "inline-block";
-    } else {
-      closeModalEl.style.display = "none";
-    }
-  }
-}
-
-  // Navigation buttons for the popup
-  if (prevTraitEl && nextTraitEl) {
-    prevTraitEl.addEventListener("click", () => {
-      if (currentSelectionIndex > 0) {
-        currentSelectionIndex--;
-        showExtraSelection();
-      }
-    });
-
-    nextTraitEl.addEventListener("click", () => {
-      if (currentSelectionIndex < extraSelections.length - 1) {
-        currentSelectionIndex++;
-        showExtraSelection();
-      }
-    });
-  }
-
-  if (closeModalEl) {
-    closeModalEl.addEventListener("click", () => {
-      console.log("🔄 Chiusura pop-up e aggiornamento UI...");
-      const raceModal = document.getElementById("raceExtrasModal");
-      if (raceModal) raceModal.classList.add('hidden');
-      sessionStorage.removeItem("popupOpened");
-
-      // ✅ Salviamo le selezioni extra PRIMA di eventuali refresh
-      saveSelectedData();
-      console.log("📝 Selezioni salvate prima dell'update:", selectedData);
-
-      if (extraModalContext === "race") {
-        showStep("step3");
-
-        setTimeout(() => {
-          console.log("🛠 Eseguo displayRaceTraits()...");
-          displayRaceTraits();
-
-          // 🔥 **Aspettiamo che `displayRaceTraits()` finisca e poi forziamo le selezioni extra**
-          setTimeout(() => {
-            console.log("✅ Forzando updateExtraSelectionsView()...");
-            updateExtraSelectionsView();
-          }, 500); // 🔥 Ritardo di 500ms per essere sicuri che il rendering sia completato
-        }, 300);
-      } else if (extraModalContext === "class") {
-        renderClassFeatures();
-      }
-    });
-  }
 
 updateExtraSelectionsView();
 
@@ -1773,6 +1160,8 @@ export {
   updateSubclasses,
   renderClassFeatures,
   openExtrasModal,
+  updateExtraSelectionsView,
+  showExtraSelection,
   displayRaceTraits,
   getTakenProficiencies,
   generateFinalJson,
